Skip unnecessary recomputations during tree search
During the tree search, positions are stored as sequences of moves from the starting position for the sake of memory efficiency. To reproduce a position in memory requires making all of those moves. Before, all of these planes were recomputed for every move. This commit gives a function in utils which does not redo the computation and uses this function to speed up the tree search.